// frontend/src/components/data-table/types.ts
import type { ColumnDef } from '@tanstack/vue-table'

export interface FacetedFilterOption {
    label: string
    value: string
    icon?: Component
}

export interface ServerPagination {
    page: number
    pageSize: number
    total: number
    onPageChange: (page: number) => void
    onPageSizeChange: (pageSize: number) => void
}

export interface DataTableProps<T> {
    loading?: boolean
    columns: ColumnDef<T, any>[]
    data: T[]
    serverPagination?: ServerPagination
}